# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/drivers:
    get:
      operationId: listAdminDrivers
      tags: [admin]
      summary: List drivers filtered by status
      description: |
        Returns driver application records for admin review.
        Requires admin authentication (Bearer token + ADMIN_EMAILS check).
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            default: pending_approval
          description: Filter by driver status
      responses:
        "200":
          description: Driver list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDriverRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/drivers/{driverId}/approve:
    post:
      operationId: approveDriver
      tags: [admin]
      summary: Approve a driver application
      description: Sets the driver's status to active.
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Driver approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/drivers/{driverId}/reject:
    post:
      operationId: rejectDriver
      tags: [admin]
      summary: Reject a driver application
      description: |
        Sets the driver's status to inactive and emails the driver with the
        provided reason explaining why their application was rejected.
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectDriverRequest"
      responses:
        "200":
          description: Driver rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionResult"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/drivers/{driverId}/reject-documents:
    post:
      operationId: rejectDriverDocuments
      tags: [admin]
      summary: Request document resubmission from a driver
      description: |
        Flags the driver's documents as rejected and records a reason.
        The driver's status remains pending_approval so they can re-upload.
        The driver sees a prominent rejection notice on their pending screen.
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectDocumentsRequest"
      responses:
        "200":
          description: Documents flagged for resubmission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionResult"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/drivers/{driverId}/clear-document-rejection:
    post:
      operationId: clearDriverDocumentRejection
      tags: [admin]
      summary: Clear a document rejection flag
      description: Removes the rejection reason from a driver record (e.g. after manually verifying re-uploaded documents).
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rejection flag cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/rides:
    get:
      operationId: listAdminRides
      tags: [admin]
      summary: List rides filtered by status
      description: Returns ride records for admin review. Requires admin authentication.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: "Filter by ride status. If omitted, returns all non-terminal rides."
      responses:
        "200":
          description: Ride list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminRideRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/rides/{rideId}/cancel:
    post:
      operationId: adminCancelRide
      tags: [admin]
      summary: Cancel a ride (admin)
      description: |
        Cancels a ride in any non-terminal status and notifies the connected
        driver via Supabase Realtime. Protected by admin authentication.
      parameters:
        - name: rideId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCancelRideRequest"
      responses:
        "200":
          description: Ride cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCancelRideResult"
        "400":
          description: Ride is already in a terminal state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ride not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AdminDriverRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        licenseDocumentPath:
          type: ["string", "null"]
        insuranceDocumentPath:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        backgroundCheckPassed:
          type: boolean
        canDriveMemberVehicle:
          type: boolean
        canDoRideshare:
          type: boolean
        canDoDelivery:
          type: boolean
        totalRidesCompleted:
          type: integer
        averageRating:
          type: number
          format: float
        documentRejectionReason:
          type: ["string", "null"]
          description: Non-null when an admin has flagged documents for resubmission.
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - status
        - backgroundCheckPassed
        - canDriveMemberVehicle
        - canDoRideshare
        - canDoDelivery
        - totalRidesCompleted
        - averageRating
    RejectDocumentsRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
          description: Human-readable reason shown to the driver explaining why their documents were rejected.
      required:
        - reason
    RejectDriverRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
          description: Human-readable reason emailed to the driver explaining why their application was rejected.
      required:
        - reason
...
    AdminActionResult:
      type: object
      properties:
        success:
          type: boolean
        driverId:
          type: string
          format: uuid
        status:
          type: string
      required:
        - success
        - driverId
        - status
    AdminRideRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        scenario:
          type: string
        tier:
          type: string
        serviceType:
          type: string
          nullable: true
        packageDescription:
          type: string
          nullable: true
        status:
          type: string
        memberId:
          type: string
          nullable: true
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        estimatedFare:
          type: number
        actualFare:
          type: number
          nullable: true
        estimatedDistanceMiles:
          type: number
        createdAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - scenario
        - tier
        - status
        - pickupAddress
        - dropoffAddress
        - estimatedFare
        - estimatedDistanceMiles
    AdminCancelRideRequest:
      type: object
      properties:
        reason:
          type: string
          description: Optional reason for the cancellation (logged server-side).
    AdminCancelRideResult:
      type: object
      properties:
        success:
          type: boolean
        rideId:
          type: string
          format: uuid
        driversNotified:
          type: integer
      required:
        - success
        - rideId
        - driversNotified
...