# lib/api-zod/src/generated/types/adminDriverRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminDriverRecord {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  status: string;
  /** @nullable */
  licenseDocumentPath?: string | null;
  /** @nullable */
  insuranceDocumentPath?: string | null;
  /** @nullable */
  profilePhotoUrl?: string | null;
  backgroundCheckPassed: boolean;
  canDriveMemberVehicle: boolean;
  canDoRideshare: boolean;
  canDoDelivery: boolean;
  totalRidesCompleted: number;
  averageRating: number;
  /**
   * Non-null when an admin has flagged documents for resubmission.
   * @nullable
   */
  documentRejectionReason?: string | null;
  /** @nullable */
  createdAt?: string | null;
}