# lib/api-zod/src/generated/types/adminRideRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminRideRecord {
  id: string;
  scenario: string;
  tier: string;
  serviceType?: string | null;
  packageDescription?: string | null;
  status: string;
  memberId?: string | null;
  pickupAddress: string;
  dropoffAddress: string;
  estimatedFare: number;
  actualFare?: number | null;
  estimatedDistanceMiles: number;
  createdAt?: Date | null;
  startedAt?: Date | null;
}